任意方向扫动都可以标雷，以及3D touch标雷
添加了两个自定义的GestureRecognizer。
将所有的代码文件按Group进行了重新整理。
一些小的拼写修正

+++ DigitView.m
@@ -82,27 +82,27 @@
 
 - (void)setDigit:(NSInteger)newDigit {
     digit = newDigit;
-    digitLabel.text = [NSString stringWithFormat:@"%lu",(unsigned long)digit];
+    digitLabel.text = [NSString stringWithFormat:@"%ld",(long)digit];
     if (digit > 999) {
         digitLabel.text = [NSString stringWithFormat:@"999"];
         digitLabel.textColor = [UIColor redColor];
     } else if (digit > 99) {
-        digitLabel.text = [NSString stringWithFormat:@"%lu",digit];
+        digitLabel.text = [NSString stringWithFormat:@"%ld",(long)digit];
         digitLabel.textColor = [UIColor redColor];
     } else if (digit > 9) {
-        digitLabel.text = [NSString stringWithFormat:@"0%lu",digit];
+        digitLabel.text = [NSString stringWithFormat:@"0%ld",(long)digit];
         digitLabel.textColor = [UIColor redColor];
     } else if (digit >= 0) {
-        digitLabel.text = [NSString stringWithFormat:@"00%lu",digit];
+        digitLabel.text = [NSString stringWithFormat:@"00%ld",(long)digit];
         digitLabel.textColor = [UIColor redColor];
     } else if (digit > -10) {
-        digitLabel.text = [NSString stringWithFormat:@"00%lu", -digit];
+        digitLabel.text = [NSString stringWithFormat:@"00%ld", (long)-digit];
         digitLabel.textColor = [UIColor yellowColor];
     } else if (digit > -100) {
-        digitLabel.text = [NSString stringWithFormat:@"0%lu", -digit];
+        digitLabel.text = [NSString stringWithFormat:@"0%ld", (long)-digit];
         digitLabel.textColor = [UIColor yellowColor];
     } else if (digit > -1000) {
-        digitLabel.text = [NSString stringWithFormat:@"%lu", -digit];
+        digitLabel.text = [NSString stringWithFormat:@"%ld", (long)-digit];
         digitLabel.textColor = [UIColor yellowColor];
     } else {
         digitLabel.text = [NSString stringWithFormat:@"999"];
